Modify res data and use params

+++ src/pages/EditEval.js
@@ -1,7 +1,7 @@
 import Header from "../components/Header";
 import Button from "../components/Button";
 import Editor from "../components/Editor";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import { useState } from "react";
 import axios from "axios";
 import { useEffect } from "react";
@@ -10,20 +10,18 @@ const SERVER_URL = `https://mapdagu.site/api/evaluations`;
 
 const EditEval = ({accessToken}) => {    
     const navigate = useNavigate();
-    //임시
-    const id =4;
-    const [newData, setNewData] = useState({
+    const { id } = useParams();
+    const [data, setData] = useState({
         name: "",
         imageNum: 0,
         score: 0,
     });
 
-    useEffect(() => {
+    useEffect(() => { 
         try{
             axios.get([SERVER_URL, id].join("/"), {headers: {Authorization: accessToken}})
                 .then(res => {
-                    console.log(res.data);
-                    setNewData({
+                    setData({
                         ...res.data,
                     });
                 })
@@ -38,32 +36,34 @@ const EditEval = ({accessToken}) => {
     }
 
     const onSubmit = async(data) => {
-      if(window.confirm("평가를 정말 수정할까요?")){
-        const score = data.score;
-        try{
-            await axios.patch([SERVER_URL, id].join("/"), {score}, {headers: {Authorization: accessToken}});
-            navigate(`/evaluate`, {replace:true});        
-        } catch(error){
-            alert(error.response.data.message);
-        }
-            // onUpdate(id, date, itemName, selectionId);
+        if(window.confirm("평가를 정말 수정할까요?")){
+            const score = data.score;
+            try{
+                await axios.patch([SERVER_URL, id].join("/"), {score}, {headers: {Authorization: accessToken}});
+                navigate(`/evaluate`, {replace:true});        
+            } catch(error){
+                alert(error.response.data.message);
+            }
         }
     }
 
     const onClickDelete = async() => {
         if(window.confirm("평가를 정말 삭제할까요?")){
-            try{
-                await axios.delete([SERVER_URL, id].join("/"), {headers: {Authorization: accessToken}});
-                navigate(`/evaluate`, {replace:true});
-            
-            } catch(error){
-                alert(error.response.data.message);
+            if(data.imageNum >= 1 && data.imageNum <=3 ){
+                alert("테스트 결과는 삭제할 수 없습니다.");
+            } else {
+                try{
+                    await axios.delete([SERVER_URL, id].join("/"), {headers: {Authorization: accessToken}});
+                    navigate(`/evaluate`, {replace:true});
+                
+                } catch(error){
+                    alert(error.response.data.message);
+                }
             }
-        //     onDelete(id);
         }
     }
     
-    if(!newData){
+    if(!data){
         return (<div>항목을 불러오고 있습니다...</div>)
     }else{
         return(
@@ -73,7 +73,7 @@ const EditEval = ({accessToken}) => {
                     leftChild={<Button text="←" onClick={handleOnBack}/>}
                     rightChild={<Button text="삭제하기" onClick={onClickDelete}/>}
                 />
-                <Editor initData={newData} onSubmit={onSubmit}/>
+                <Editor initData={data} onSubmit={onSubmit}/>
             </div>
         );
     }
